refactor: move pkg/cloudlogin to cmd/kubectl-testkube/cloudlogin
