+++ cmd/kubectl-testkube/commands/common/oauth.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"fmt"
 
+	"github.com/kubeshop/testkube/cmd/kubectl-testkube/cloudlogin"
 	"github.com/kubeshop/testkube/cmd/kubectl-testkube/config"
-	"github.com/kubeshop/testkube/pkg/cloudlogin"
 )
 
 // GetOAuthAccessToken strictly checks for OAuth authentication and returns the access token
